requirements.txt
flask==2.3.3
python-dotenv==1.0.0
davia==0.1.0
pandas==2.1.0
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.2
black==23.7.0
flake8==6.1.0
python-dateutil==2.8.2
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7 